+++ sentry-ruby/lib/sentry-ruby.rb
@@ -83,7 +83,8 @@ module Sentry
     attr_reader :backpressure_monitor
 
     # @!attribute [r] metrics_aggregator
-    #   @return [Metrics::Aggregator, nil]
+    # @deprecated
+    # @return [nil]
     attr_reader :metrics_aggregator
 
     ##### Patch Registration #####
@@ -241,7 +242,7 @@ module Sentry
       @background_worker = Sentry::BackgroundWorker.new(config)
       @session_flusher = config.session_tracking? ? Sentry::SessionFlusher.new(config, client) : nil
       @backpressure_monitor = config.enable_backpressure_handling ? Sentry::BackpressureMonitor.new(config, client) : nil
-      @metrics_aggregator = config.metrics.enabled ? Sentry::Metrics::Aggregator.new(config, client) : nil
+      @metrics_aggregator = nil
       exception_locals_tp.enable if config.include_local_variables
       at_exit { close }
     end
@@ -260,12 +261,6 @@ module Sentry
       if @backpressure_monitor
         @backpressure_monitor.kill
         @backpressure_monitor = nil
-      end
-
-      if @metrics_aggregator
-        @metrics_aggregator.flush(force: true)
-        @metrics_aggregator.kill
-        @metrics_aggregator = nil
       end
 
       if client = get_current_client