+++ sentry-ruby/spec/sentry/envelope/item_spec.rb
@@ -12,8 +12,6 @@ RSpec.describe Sentry::Envelope::Item do
       ['span', 'span'],
       ['profile', 'profile'],
       ['check_in', 'monitor'],
-      ['statsd', 'metric_bucket'],
-      ['metric_meta', 'metric_bucket'],
       ['event', 'error'],
       ['client_report', 'internal'],
       ['unknown', 'default']